+++ api/src/paths/observation/index.ts
@@ -2,6 +2,7 @@ import { Request, RequestHandler } from 'express';
 import { Operation } from 'express-openapi';
 import { SYSTEM_ROLE } from '../../constants/roles';
 import { getDBConnection } from '../../database/db';
+import { getSystemUserFromRequest } from '../../express/request';
 import { IObservationAdvancedFilters } from '../../models/observation-view';
 import { observervationsWithSubcountDataSchema } from '../../openapi/schemas/observation';
 import { paginationRequestQueryParamSchema } from '../../openapi/schemas/pagination';
@@ -13,7 +14,6 @@ import {
   makePaginationOptionsFromRequest,
   makePaginationResponse
 } from '../../utils/pagination';
-import { getSystemUserFromRequest } from '../../utils/request';
 
 const defaultLog = getLogger('paths/observation/index');
